Resolve the cherry-pick of #108464 onto 26.3. Exactly one file conflicts,
`src/Functions/hasColumnInTable.cpp`, in two marker regions.

#110881 removed the `hostname[, username[, password]]` overload from master two
days before #108464 merged, so the merged form of #108464 has a single code
path. 26.3 predates that removal and still dispatches on `host_name.empty()`.

Resolution:

  * includes: union of both sides - the two `Access/Common` headers plus the
    branch's `<Core/Settings.h>`, which 26.3 still needs for `getSettingsRef`.
  * `executeImpl`: the `checkAccess(AccessType::SHOW_COLUMNS, ...)` call is
    hoisted ABOVE the branch's `if (host_name.empty())` dispatch, which is kept
    verbatim. It must not go inside the local arm: on 26.3 the remote arm
    otherwise connects as `default` before any privilege is examined, and the
    check `getStructureOfRemoteTableInShard` already performs for local shards
    is a no-op today because `create` passed `getGlobalContext`, which has full
    access by construction. Passing the query context makes that check
    effective too.
  * the `:::note Privilege SHOW COLUMNS is required` block in
    `REGISTER_FUNCTION(HasColumnInTable)` merges on its own and is preserved
    alongside the branch's own variadic `Syntax` string and `Arguments`
    entries. Taking either side of the file wholesale would have dropped it.

The two test files are taken from `72a9159bda6a`, the pre-reconciliation
revision of #108464, rather than from its merged form. The merged test was
reconciled for the same reason the file conflicts - master had lost the
overload - so it cannot exercise the remote path that this change gates on
26.3. The pre-reconciliation test is a strict superset and covers the
`localhost` and non-loopback `127.0.0.2` overloads.

Resolution independently reproduced from the recipe in
https://github.com/ClickHouse/ClickHouse/pull/120196#issuecomment-3298772483
tree `50b4bd9685d21573ceb370d3b7c0e27f45a7a81c`, resolved blob
`9be8cf74c949894a97b052056bfd351b32b0f1ec`; both match. The resolved
translation unit type-checks against the 26.3 tree.

+++ src/DataTypes/Native.cpp
@@ -391,7 +391,7 @@ llvm::Constant * getColumnNativeValue(llvm::IRBuilderBase & builder, const DataT
 
         if constexpr (std::is_floating_point_v<T>)
         {
-            return llvm::ConstantFP::get(type, static_cast<double>(element));
+            return llvm::ConstantFP::get(type, element);
         }
         else if constexpr (is_integer<T>)
         {
